Use `ssize_t` for `ignore`

+++ src/lib/arch/unix/ArchNetworkBSD.cpp
@@ -316,7 +316,7 @@ int ArchNetworkBSD::pollSocket(PollEntry pe[], int num, double timeout) {
   if (n > 0 && unblockPipe != nullptr && (pfd[num].revents & POLLIN) != 0) {
     // the unblock event was signalled.  flush the pipe.
     char dummy[100];
-    int ignore;
+    ssize_t ignore;
 
     do {
       ignore = m_deps.read(unblockPipe[0], dummy, sizeof(dummy));
